input-chooser: Simplify marking language or country layouts non-extra
choose_non_extras() was previously called after adding the layouts for
the locale, language, and country, but before adding all other layouts
from xkb and ibus. So by construction only the first of those layouts
are marked as non-extra.

We can simplify this by explicitly passing an additional is_extra
parameter, setting this to TRUE for those layouts, and flipping it back
to FALSE once the MIN_ROWS threshold is reached.

https://phabricator.endlessm.com/T35040