timezone: Hide when automatically detected
In downstream Endless, the timezone page shouldn't be shown if a timezone
has been automatically detected.
Talking to the GNOME Design team, this behavior would also be desirable
but unfortunately it is not clear yet whether a detected timezone will
be correct in every territory.

The argument for skipping the page if location has been detected in some
Endless images is that this makes the FBE process nicer enough to risk
sometimes silently selecting the wrong timezone. If we believe this for
some users, without any data about how often the wrong timezone is
selected, then we should apply this behaviour for all systems.

https://phabricator.endlessm.com/T23631