language-chooser: Expand by default
In early 2017, we found that users do not associate the ⋮ icon with
“show more”:

> This leads some users (including our Romanian translator, who knows
> for a fact we do support Romanian!) to conclude we don't support any
> other languages.

It is tricky to label the expander any better because, on the language
screen, we do not know what languages the user can read, so we are
limited to iconography.

Expand the list by default. Together with our setting
org.gnome.desktop.interface overlay-scrolling to false to make
scrollbars always-visbile, this presents a clear hint that you can
scroll down to see more languages. There is also a prominent search box.

That report also says:

> We only show "English (United States)" in the first screen of the FBE
> on a "base" image […]

which is no longer the case, so it would be interesting to retest this
with the current presentation that shows a short list of widely-spoken
languages.

(Rebase 43: Previous versions of this patch made other changes to this
page, but since the GTK 4 rewrite these have become unnecessary.)

https://phabricator.endlessm.com/T15310

https://phabricator.endlessm.com/T35040