fixup! page-util: Add API to read the EOS image version
Being able to pass a number of different paths in here is a remnant of
our split disk configuration. Nowadays we only have a single disk.

I think we check both /sysroot and / for the benefit of "converted"
systems, which are now extremely rare, but it's easy enough to support
both in one place.

https://phabricator.endlessm.com/T35040