fixup! timezone: set default 12h/24h from locale
The signature of update_timezone() is:

    static void
    update_timezone (GisTimezonePage *page, TzLocation *location)

However a callback to the GObject::notify signal should accept 3
arguments:

    static void
    notify_cb (GObject    *object,
               GParamSpec *param_spec,
               gpointer    user_data)

So if this callback were to be called, the 'page' argument would in fact
be a GSettings * but would be treated as a GisTimezonePage *, while the
'location' argument would in fact be a GParamSpec * but treated as a
TzLocation *.

Happily this callback is never called, due to a second bug. The 'notify'
signal is emitted when a GObject property's value may have changed. The
"x::y" syntax matches signal "x" with detail string "y". For
GObject::notify, the detail is the name of the property whose value may
have changed. GSettings has no property named clock-format, so "notify"
is never emitted with details "clock-format", so the callback is never
called.

The likely intention was to connect to the GSettings::changed signal,
which takes as detail the name of the setting which may have changed.
But there is no need to monitor this because we already monitor the
GnomeWallClock:clock property for changes, and GnomeWallClock emits
notify for this property when the clock-format gsettting is changed.

So, remove the doubly-incorrect signal connection without replacement.

https://phabricator.endlessm.com/T35040