factory dialog: Show machine ID
This is useful in a factory context to check that two machines have not
been incorrectly cloned from a disk image that had already been booted.

Make gis_factory_dialog_new() private: it is not called outside
gis-factory-dialog.c, and this saves changing its signature twice.

GLib and friends do not provide an API function to read the machine ID.
GLib has an internal function _g_dbus_get_machine_id(), but this is not
exposed.

libsystemd provides sd_id128_get_machine(), but this returns an
sd_id12_t which we would immediately turn back into a string for
display, and we don't currently link to libsystemd.

Reading the file is a one-liner and we don't need to worry about any
complicated error handling. The hardest part is remembering to remove
the trailing newline.

https://phabricator.endlessm.com/T35570